Revert "feat: [퀘스트 - 촉각] TextureButton 새로운 파일로 분류"
This reverts commit 95b977207c7d50b4a8bc869185be2937001452bd.
